refactor: change context provider
